extract json encoder and decorder as JSONHandler
